+++ Assets/Code/Scripts/Manager/GameManager.cs
@@ -12,6 +12,7 @@ public class GameManager : MonoSingleton<GameManager> // ½Ì±ÛÅæ »ç¿ë
 
     [Header("ÄÁÆ®·Ñ·¯")]
     public PlayerController playerController;
+    public GrapplingHook grapplingHook;
 
     [Header("½ºÅÈ")]
     public PlayerStats playerStats;
