--- Assets/Code/Scripts/Player/SwingBoostController.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: 11ed0b504517e004eb8510cda5abb9b7
