--- Assets/Code/Scripts/Player/SwingGaugeController.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: 999c92b5ce3397b499191207a11e1ed1
